+++ src/js/App.js
@@ -4,9 +4,9 @@ import Home from './components/home/Home.js';
 import Login from './components/login/Login.js';
 import SignUp from './components/signup/Signup.js';
 import Header from './components/header/Header.js';
-import Stations from './components/stations/Stations.js';
-import Lines from './components/lines/Lines.js';
-import Sections from './components/sections/Sections.js';
+import Station from './components/station/Station.js';
+import Line from './components/line/Line.js';
+import Section from './components/section/Section.js';
 
 import { showSnackbar } from './utils/snackbar.js';
 import { getLocalStorageItem } from './utils/storage.js';
@@ -72,9 +72,9 @@ class App {
       switchURL: this.switchURL.bind(this),
       showSnackbar: this.#showSnackbar.bind(this),
     });
-    this.stations = new Stations();
-    this.sections = new Sections();
-    this.lines = new Lines();
+    this.stations = new Station();
+    this.sections = new Section();
+    this.lines = new Line();
   }
 
   _registerRouterComponent() {
